% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{generates hclust from data}
\usage{
cluster_data(
  data,
  similarity_method = c("basepair", "basepair_sorensen", "basepair_max-union", "percent",
    "percent_sorensen", "percent_max-union"),
  dist_method = "euclidean",
  clust_method = "complete",
  name = NULL
)
}
\arguments{
\item{data}{an object}

\item{similarity_method}{one of 'sorensen', 'basepair',
or 'counts'.}

\item{dist_method}{method to provide to \code{stats::dist}}

\item{clust_method}{method to provide to \code{stats:hclust}}
}
\description{
generates hclust from data
}
