% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_vctr_impl.R
\name{descendants}
\alias{descendants}
\title{reports the descendants of each node according
to the current encoding of the tree_vctr. For
nodes that are leaves, this should be identical
to the node_id() funciton.}
\usage{
descendants(x)
}
\description{
reports the descendants of each node according
to the current encoding of the tree_vctr. For
nodes that are leaves, this should be identical
to the node_id() funciton.
}
