% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_cluster.R
\name{df_cluster}
\alias{df_cluster}
\title{cluster data frame}
\usage{
df_cluster(
  data,
  id_rows,
  id_cols,
  values_from,
  rows_to = "row_tree",
  cols_to = "col_tree",
  ...,
  .linkage_method = "complete",
  .distance_method = "euclidean",
  .p = 2,
  cluster_row = TRUE,
  cluster_col = TRUE,
  by_rows = NULL,
  by_cols = NULL,
  use_labels = FALSE
)
}
\description{
cluster data frame
}
\examples{
mat <- matrix(0, 6, 5)
mat[1:3, 1:3] <- rnorm(9, 10)
mat[1:2, 4:5] <- rnorm(4, 4)
mat[4:6, 1:2] <- rnorm(6, 7)
mat[4:6, 3:5] <- rnorm(9, 2)
mat[3, 4:5] <- rnorm(2, 15)
dimnames(mat) <- list(sprintf("rows\%i", 1:6), sprintf("cols\%i", 1:5))
df <- tibble::as_tibble(mat, rownames = "rows") |>
  dplyr::mutate(grow = rep(1:2, each = 3)) |>
  tidyr::pivot_longer(cols = -c(rows, grow), names_to = "cols") |>
  dplyr::mutate(gcol = rep(c(1, 1, 1, 2, 2), 6))
}
