% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_inner.R
\name{generate_inner_slice}
\alias{generate_inner_slice}
\title{creates a data.frame that maps inner_nodes
to the indices of the input tree_vctr. Note
that only unique values of tree_vctr are
considered. If leaf nodes are missing in the
input, then they will not appear in the output.}
\usage{
generate_inner_slice(tree_vec, node_level = NULL, only_inner = TRUE)
}
\description{
creates a data.frame that maps inner_nodes
to the indices of the input tree_vctr. Note
that only unique values of tree_vctr are
considered. If leaf nodes are missing in the
input, then they will not appear in the output.
}
