% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_vctr.R
\name{new_tree_vctr}
\alias{new_tree_vctr}
\title{a tree_vctr object}
\usage{
new_tree_vctr(node = integer(), which_tree = integer(), tree)
}
\arguments{
\item{node}{an integer vector represeting some encoded node id}

\item{which_tree}{identifies which element of \code{node} maps to which hclust
tree in the \code{tree} attribute}

\item{tree}{a list of \code{hclust} objects}
}
\description{
This object allows a way to encode many heirachical objects along with
the rows of a matrix-like object. Each element of a \code{tree_vctr} object
will have a node id and a tree id, which together identify a specific
observation within the \code{hclust} object stored in the \code{tree} attribute.
DEFUNCT: The node id may be encoded in two ways, which mainly provide convience
with sorting this vector according to either the order of observations
or an order that would cluster the observations per hclust object.
}
