% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_classes.R
\name{simple_range}
\alias{simple_range}
\title{a light weight way to constuct a "Range" class
when we only care about the start and stop positions
since splitting a GRangesList can be fairly expensive}
\usage{
simple_range(x)
}
\description{
a light weight way to constuct a "Range" class
when we only care about the start and stop positions
since splitting a GRangesList can be fairly expensive
}
