% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_classes.R
\docType{data}
\name{supported_classes}
\alias{supported_classes}
\alias{GR}
\alias{GRL}
\alias{SE}
\alias{RSE}
\title{Supported Classes}
\format{
An object of class \code{classRepresentation} of length 1.

An object of class \code{classRepresentation} of length 1.

An object of class \code{classRepresentation} of length 1.

An object of class \code{classRepresentation} of length 1.
}
\usage{
GR

GRL

SE

RSE
}
\description{
Possible classes to dispatch on
}
\keyword{datasets}
