% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeClustDemo-package.R
\docType{package}
\name{treeClustDemo-package}
\alias{treeClustDemo}
\alias{treeClustDemo-package}
\title{treeClustDemo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
